feat: add insertParcelHandlingInfo() request feat: make insertParcelHandlingInfo() execute after saving Parcel into db feat: change findShelfbyVodkaId() to findShelfIdByVodkaId()

+++ clients/src/main/java/com/rszumlas/clients/vodka/VodkaRequest.java
@@ -3,6 +3,6 @@ package com.rszumlas.clients.vodka;
 public record VodkaRequest(
         String name,
         Integer bottle_size,
-        Integer voltage
+        Double voltage
 ) {
 }
